Break requirements.txt into environment specific modules
Our previous approach to dependencies was creating a high barrier to
entry for new contributors (and new Pythonistas/Djangonauts) since
it required them to have a production environment installed locally
just to install the requirements files.

This change moves production-specific dependencies to a separate
module, simplifies local setup, and opens the possiblity of having
development specific dependencies in the future.

closes #37