Add a CONTRIBUTING.md file and simplify README
Now that the requirements files are split up in #37 and the local dev
environment has been simplified, we can now add a detailed and welcoming
contribution guide.

This absorbs the install instructions from the README and expounds on
the install and local development process so as to enable new developers
to contribute to the project easly.

There were no guidelines in the README as far as commit hygene or other
contribution guidelines, but it wouldn't be difficult to add later.

fixes: #19, #26, #27
relies on: #37